browser-bot-master/requirements.txt
playwright>=1.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
apscheduler>=3.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0
